feat: centralize duplicated `OfficialCollection` and `ExternalLink` components
Adds a new `components/` directory to the view models directory which is intended to house shared components. Refactors the official collection component into the `collections` view model folder, which is more appropriate given the content.
