+++ src/mavedb/view_models/score_set.py
@@ -19,6 +19,8 @@ from mavedb.models.enums.mapping_state import MappingState
 from mavedb.models.enums.processing_state import ProcessingState
 from mavedb.view_models import record_type_validator, set_record_type
 from mavedb.view_models.base.base import BaseModel
+from mavedb.view_models.collection import OfficialCollection
+from mavedb.view_models.components.external_link import ExternalLink
 from mavedb.view_models.contributor import Contributor, ContributorCreate
 from mavedb.view_models.doi_identifier import (
     DoiIdentifier,
@@ -47,20 +49,6 @@ from mavedb.view_models.user import SavedUser, User
 from mavedb.view_models.utils import all_fields_optional_model
 
 UnboundedRange = tuple[Union[float, None], Union[float, None]]
-
-
-class ExternalLink(BaseModel):
-    url: Optional[str] = None
-
-
-class OfficialCollection(BaseModel):
-    badge_name: str
-    name: str
-    urn: str
-
-    class Config:
-        arbitrary_types_allowed = True
-        from_attributes = True
 
 
 class ScoreSetBase(BaseModel):
